Fix single complete file transfer from importing all files
When importing multiple files at once, each completed file will trigger inotifywatch, and the script will attempt to import all files, even the ones unfinished. To fix this, the scripts have to individually process one file at a time. This includes a refactor and major change of the python code