Added the following CW PRs:
#3355 - Shows the number of items in each shelf in the Web UI
#3352 - Kobo sync's logic is incorrect, and looks like it pretty much always has been. This is seen on initial sync of a freshly reset device against CW configured to sync all shelves. With 1500+ books to sync, CW syncs only about 250 or so before saying it's done. That's two full batches of 100, followed by a partial batch.

The problem is in the way the query performs paging. On the first pass it discovers all the books that need to be synced. On the second pass it does pretty well. Third pass, it does badly. All this is because of the way it currently detects which books have been modified since last sync. As it goes through the sync, this query result set gets progressively smaller, incorrectly.

Attempt to repair the logic, and seems to now work perfectly, working a lot better now for both initial sync and maintenance sync